server/requirements.txt
Flask
kafka-python
redis
torch
torchvision
